Moved production to constants
