+++ app/modal-windows/main.ts
@@ -9,7 +9,7 @@ import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
 import { enableProdMode } from '@angular/core';
 
 import { ModalWindowModule } from "./modal-window.module";
-import { IS_PRODUCTION } from "../main";
+import { IS_PRODUCTION } from "../constants";
 
 import "file-loader?name=modal.html!extract-loader!./index.html";
